[lldb][NFC] Expose GetReflectionContext in SwiftLanguageRuntime's API
This is in preparation for building type infos from DWARF, which will
need access to the reflection context and it's builder.
